requirements.txt
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
bce-python-sdk==0.8.79
boto3==1.26.109
botocore==1.29.109
brotlipy==0.7.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
datasets==2.6.0
decorator==5.1.1
dill==0.3.4
entrypoints==0.4
evaluate==0.4.0
fastapi==0.93.0
ffmpy==0.3.0
filelock==3.9.0
Flask==2.2.3
Flask-Babel==2.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gradio==3.24.1
gradio_client==0.0.8
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.12.2
openai==0.27.2
opt-einsum==3.3.0
orjson==3.8.9
packaging==23.0
paddle-bfloat==0.1.7
paddle2onnx==1.0.5
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle==2.4.2
pandas==1.5.3
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
prettytable==3.6.0
promcse==0.0.2
protobuf==3.20.0
psutil==5.9.4
pyarrow==11.0.0
pycryptodome==3.17
pydantic==1.10.6
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-pretrained-bert==0.6.2
PyYAML==6.0
responses==0.18.0
rfc3986==1.5.0
rich==13.3.2
rouge==1.0.1
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.0
semantic-version==2.10.0
sentencepiece==0.1.97
seqeval==1.2.2
sniffio==1.3.0
starlette==0.25.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.64.1
transformers==4.28.1
typer==0.7.0
uc-micro-py==1.0.1
uvicorn==0.20.0
visualdl==2.4.2
wcwidth==0.2.6
websockets==11.0.1
Werkzeug==2.2.3
xxhash==3.2.0
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
